+++ src/store/codec_conversions.js
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import { index, create, destroy, update } from "../api/codec_conversions";
-import { fetchAll } from "./commit";
+import { fetchAll } from "./actions";
 
 export default {
   namespaced: true,
